// VIS_Application/app/Model/HumanResource/EmployeeManagement/EmployeeList.ts
export interface IEmployeeList {
    mode: string;
    UserId: number;
    Usertype: string;
    Id: number;
    EmployeeCode: string; 
    CompanyName: string;
    Employee_Name: string;
    Email: string;
    Department:string;
    TotalExp: string;
    JoiningDate: Date;
    RelevingDate:Date;
    Grace: number;
    SL_CL: string;

}